feat(@angular-devkit/core): add an Annotation Map as a PoC
This would be useful for separating dependencies, in a DI kind of scheme, but
lower level (no annotations, just get/set).

Two places this can be used; to add annotation metadata to files in Schematics,
and in Architect as the context to contain more information (e.g. Workspace)
without forcing those dependencies on Architect itself.